+++ spec/requests/dao_revocations_controller_spec.rb
@@ -80,6 +80,8 @@ RSpec.describe DaoRevocationsController, type: :request do
           reason_school_rating_improved_note: "",
           reason_safeguarding_addressed: "",
           reason_safeguarding_addressed_note: "",
+          reason_change_to_policy: "",
+          reason_change_to_policy_note: "",
           minister_name: "Minister Name",
           date_of_decision: ""
         }
@@ -145,6 +147,8 @@ RSpec.describe DaoRevocationsController, type: :request do
           reason_school_rating_improved_note: "",
           reason_safeguarding_addressed: "",
           reason_safeguarding_addressed_note: "",
+          reason_change_to_policy: "",
+          reason_change_to_policy_note: "",
           minister_name: "Minister Name",
           date_of_decision: ""
         }
@@ -181,6 +185,8 @@ RSpec.describe DaoRevocationsController, type: :request do
           reason_school_rating_improved_note: nil,
           reason_safeguarding_addressed: "false",
           reason_safeguarding_addressed_note: nil,
+          reason_change_to_policy: "true",
+          reason_change_to_policy_note: "The policy changed",
           minister_name: "Minister Name"
         }
       }
@@ -201,6 +207,8 @@ RSpec.describe DaoRevocationsController, type: :request do
           reason_school_rating_improved_note: "",
           reason_safeguarding_addressed: "false",
           reason_safeguarding_addressed_note: "",
+          reason_change_to_policy: "true",
+          reason_change_to_policy_note: "The policy changed",
           minister_name: "Minister Name",
           date_of_decision: "2024-01-01"
         }
@@ -215,6 +223,7 @@ RSpec.describe DaoRevocationsController, type: :request do
           reason_school_closed: "",
           reason_school_rating_improved: "",
           reason_safeguarding_addressed: "",
+          reason_change_to_policy: "",
           minister_name: "Minister Name"
         }
       }
@@ -232,11 +241,13 @@ RSpec.describe DaoRevocationsController, type: :request do
         dao_revocation_stepped_form:
         {
           reason_school_closed: "true",
-          reason_school_closed_note: "Details fo the schoool closing.",
+          reason_school_closed_note: "Details of the school closing.",
           reason_school_rating_improved: "false",
           reason_school_rating_improved_note: nil,
           reason_safeguarding_addressed: "false",
           reason_safeguarding_addressed_note: nil,
+          reason_change_to_policy: "false",
+          reason_change_to_policy_note: nil,
           minister_name: "Minister Name",
           date_of_decision: "2024-1-1"
         }